moved common fixtures to conftest.py for reuse
