+++ services/api/tests/functional/src/test_genre.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.mark.asyncio
+async def test_genres():
+    pass
